Report errors to commit requests
Further improves the commit process of ConsumerGroupStream.

The protocol was not correctly bubbling error responses up to
committers, giving a false sense of commit requests being successful.

This leads consumers to believe their offset committed, when it did not.

Primary changes are done in ConsumerGroupStream as it has a better model
for per-partition responses, but effort was made to ConsumerGroup too, although
it will give an all or nothing type response.

ConsumerGroup's behavior is dangerous for consumers anyways, so end users of
that method can check the topic response to know which partitions failed.

Format of the protocol response has changed which may be a breaking change
if anyone depended on the old format, but the old format was invalid and lost
response data for errors and can not be maintained.